pyproject.toml
[tool.poetry]
name = "myweb100"
version = "0.1.0"
description = ""
authors = ["jianjian001 <jianjian001@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
beautifulsoup4 = "^4.9.3"
PyMySQL = "^0.9.2"
aiohttp = "^3.7.4"
aiomysql = "^0.0.21"
qiniu = "^7.4.0"
aiofiles = "^0.7.0"
Flask-Babel = "^2.0.0"
feedparser = "^6.0.2"
lxml = "^4.6.3"
gevent = "^21.1.2"
gunicorn = "^20.1.0"
requests = "^2.25.1"
redis = "^3.5.3"
flask-redis = "^0.4.0"
pony = "^0.7.14"
Flask-WTF = "^0.15.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
